+++ neoforge/src/main/java/com/jahirtrap/walljump/WallJumpClient.java
@@ -1,0 +1,22 @@
+package com.jahirtrap.walljump;
+
+import net.minecraft.client.KeyMapping;
+import net.neoforged.api.distmarker.Dist;
+import net.neoforged.bus.api.SubscribeEvent;
+import net.neoforged.fml.common.EventBusSubscriber;
+import net.neoforged.neoforge.client.event.RegisterKeyMappingsEvent;
+import org.lwjgl.glfw.GLFW;
+
+@EventBusSubscriber(modid = WallJumpMod.MODID, bus = EventBusSubscriber.Bus.MOD, value = Dist.CLIENT)
+public class WallJumpClient {
+    public static KeyMapping KEY_WALL_JUMP = new KeyMapping(
+            "key.walljump.walljump",
+            GLFW.GLFW_KEY_LEFT_SHIFT,
+            "key.categories.walljump"
+    );
+
+    @SubscribeEvent
+    public static void registerKeyBinding(RegisterKeyMappingsEvent event) {
+        event.register(KEY_WALL_JUMP);
+    }
+}