+++ neoforge/src/main/java/com/jahirtrap/walljump/logic/SpeedBoostLogic.java
@@ -1,7 +1,7 @@
 package com.jahirtrap.walljump.logic;
 
-import com.jahirtrap.walljump.init.ModConfig;
 import com.jahirtrap.walljump.init.ModEnchantments;
+import com.jahirtrap.walljump.init.ServerConfig;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.player.LocalPlayer;
 import net.minecraft.core.Holder;
@@ -21,7 +21,7 @@ import net.neoforged.api.distmarker.OnlyIn;
 @OnlyIn(Dist.CLIENT)
 public class SpeedBoostLogic {
     public static void doSpeedBoost(LocalPlayer pl) {
-        if ((!ModConfig.enableEnchantments || !ModConfig.enableSpeedBoost) && (ModConfig.sprintSpeedBoost == 0 && ModConfig.elytraSpeedBoost == 0))
+        if ((!ServerConfig.enableEnchantments || !ServerConfig.enableSpeedBoost) && (ServerConfig.sprintSpeedBoost == 0 && ServerConfig.elytraSpeedBoost == 0))
             return;
         int jumpBoostLevel = 0;
         MobEffectInstance jumpBoostEffect = pl.getEffect(MobEffects.JUMP);
@@ -37,7 +37,7 @@ public class SpeedBoostLogic {
                 if (pl.getXRot() < 30f)
                     pl.setDeltaMovement(motion.subtract(motion.scale(0.05)));
             } else if (Minecraft.getInstance().options.keySprint.isDown() && pl.input.hasForwardImpulse()) {
-                float elytraSpeedBoost = (float) (ModConfig.elytraSpeedBoost + (getEquipmentBoost(pl, EquipmentSlot.CHEST) * (float) ModConfig.speedBoostMultiplier));
+                float elytraSpeedBoost = (float) ServerConfig.elytraSpeedBoost + (getEquipmentBoost(pl, EquipmentSlot.CHEST) * (float) ServerConfig.speedBoostMultiplier);
                 Vec3 boost = new Vec3(look.x, look.y, look.z).normalize().scale(elytraSpeedBoost);
                 if (motion.length() <= boost.length())
                     pl.setDeltaMovement(motion.add(boost.scale(0.05)));
@@ -45,7 +45,7 @@ public class SpeedBoostLogic {
                     pl.level().addParticle(ParticleTypes.FIREWORK, pos.x, pos.y, pos.z, 0, 0, 0);
             }
         } else if (pl.isSprinting()) {
-            float sprintSpeedBoost = (float) (ModConfig.sprintSpeedBoost + (getEquipmentBoost(pl, EquipmentSlot.FEET) * (ModConfig.speedBoostMultiplier / 2)));
+            float sprintSpeedBoost = (float) (ServerConfig.sprintSpeedBoost + (getEquipmentBoost(pl, EquipmentSlot.FEET) * (ServerConfig.speedBoostMultiplier / 2)));
             if (!pl.onGround()) sprintSpeedBoost /= 3.125F;
 
             Vec3 boost = new Vec3(look.x, 0.0, look.z).scale(sprintSpeedBoost * 0.125F);
@@ -54,7 +54,7 @@ public class SpeedBoostLogic {
     }
 
     private static int getEquipmentBoost(LocalPlayer pl, EquipmentSlot slot) {
-        if (!ModConfig.enableEnchantments || !ModConfig.enableSpeedBoost)
+        if (!ServerConfig.enableEnchantments || !ServerConfig.enableSpeedBoost)
             return 0;
         ItemStack stack = pl.getItemBySlot(slot);
         if (!stack.isEmpty()) {